Minor bugfixes.
Some refactoring of compare dialog. Bugfix for exporting misses.
