+++ .gitignore
@@ -7,6 +7,7 @@
 *.lai
 *.so
 *.so.*
+*.dll
 *.dylib
 
 # Qt-es
@@ -14,8 +15,7 @@ object_script.*.Release
 object_script.*.Debug
 *_plugin_import.cpp
 /.qmake.cache
-.qmake.stash
-*_resource.rc
+/.qmake.stash
 *.pro.user
 *.pro.user.*
 *.qbs.user
@@ -51,6 +51,7 @@ compile_commands.json
 # QtCreator local machine specific files for imported projects
 *creator.user*
 
-# Qt outputs
-*/release*
-*/debug*
+# Outputs
+*release*
+*debug*
+*_resource.rc
