+++ FileVerifier.qrc
@@ -1,0 +1,6 @@
+<!DOCTYPE RCC>
+<RCC version="1.0">
+	<qresource>
+		<file>media/FileVerifier.png</file>
+	</qresource>
+</RCC>
