--- FileVerifier/FileVerifier.qrc
@@ -1,6 +1,0 @@
-<!DOCTYPE RCC>
-<RCC version="1.0">
-	<qresource>
-		<file>FileVerifier.png</file>
-	</qresource>
-</RCC>
