--- FileVerifier/LICENSE (Icon).txt
@@ -1,1 +1,0 @@
-https://icons8.de/license
