+++ mainwindow.ui
@@ -1,0 +1,154 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ui version="4.0">
+ <class>MainWindow</class>
+ <widget class="QMainWindow" name="MainWindow">
+  <property name="geometry">
+   <rect>
+    <x>0</x>
+    <y>0</y>
+    <width>682</width>
+    <height>480</height>
+   </rect>
+  </property>
+  <widget class="QWidget" name="centralWidget">
+   <layout class="QVBoxLayout" name="verticalLayout">
+    <item>
+     <layout class="QHBoxLayout" name="horizontalLayout">
+      <item>
+       <widget class="QComboBox" name="comboBoxTables">
+        <property name="minimumSize">
+         <size>
+          <width>300</width>
+          <height>0</height>
+         </size>
+        </property>
+       </widget>
+      </item>
+     </layout>
+    </item>
+    <item>
+     <widget class="QTableView" name="tableView"/>
+    </item>
+    <item>
+     <widget class="QProgressBar" name="progressBar">
+      <property name="value">
+       <number>0</number>
+      </property>
+      <property name="textVisible">
+       <bool>false</bool>
+      </property>
+     </widget>
+    </item>
+   </layout>
+  </widget>
+  <widget class="QMenuBar" name="menuBar">
+   <property name="geometry">
+    <rect>
+     <x>0</x>
+     <y>0</y>
+     <width>682</width>
+     <height>21</height>
+    </rect>
+   </property>
+   <widget class="QMenu" name="menuDb">
+    <property name="title">
+     <string>Database</string>
+    </property>
+    <addaction name="actionDbNew"/>
+    <addaction name="actionDbOpen"/>
+    <addaction name="separator"/>
+    <addaction name="actionE_xit"/>
+   </widget>
+   <widget class="QMenu" name="menuVerify">
+    <property name="title">
+     <string>Verify</string>
+    </property>
+    <addaction name="actionVerifyFile"/>
+    <addaction name="actionVerifyDirectory"/>
+   </widget>
+   <widget class="QMenu" name="menuExport">
+    <property name="title">
+     <string>Export</string>
+    </property>
+    <addaction name="actionCurrent_Table"/>
+   </widget>
+   <widget class="QMenu" name="menuHash">
+    <property name="title">
+     <string>Hash</string>
+    </property>
+    <addaction name="actionHashFile"/>
+    <addaction name="actionHashDirectory"/>
+   </widget>
+   <widget class="QMenu" name="menuHelp">
+    <property name="title">
+     <string>Help</string>
+    </property>
+    <addaction name="actionAbout"/>
+    <addaction name="actionAbout_Qt"/>
+   </widget>
+   <addaction name="menuDb"/>
+   <addaction name="menuHash"/>
+   <addaction name="menuVerify"/>
+   <addaction name="menuExport"/>
+   <addaction name="menuHelp"/>
+  </widget>
+  <action name="actionHashFile">
+   <property name="text">
+    <string>File(s)</string>
+   </property>
+  </action>
+  <action name="actionHashDirectory">
+   <property name="text">
+    <string>Directory</string>
+   </property>
+  </action>
+  <action name="actionSet_file_filter">
+   <property name="text">
+    <string>Set file filter</string>
+   </property>
+  </action>
+  <action name="actionVerifyFile">
+   <property name="text">
+    <string>File(s)</string>
+   </property>
+  </action>
+  <action name="actionVerifyDirectory">
+   <property name="text">
+    <string>Directory</string>
+   </property>
+  </action>
+  <action name="actionCurrent_Table">
+   <property name="text">
+    <string>Current Table</string>
+   </property>
+  </action>
+  <action name="actionDbNew">
+   <property name="text">
+    <string>New...</string>
+   </property>
+  </action>
+  <action name="actionDbOpen">
+   <property name="text">
+    <string>Open...</string>
+   </property>
+  </action>
+  <action name="actionE_xit">
+   <property name="text">
+    <string>E&amp;xit</string>
+   </property>
+  </action>
+  <action name="actionAbout">
+   <property name="text">
+    <string>About</string>
+   </property>
+  </action>
+  <action name="actionAbout_Qt">
+   <property name="text">
+    <string>About Qt</string>
+   </property>
+  </action>
+ </widget>
+ <layoutdefault spacing="6" margin="11"/>
+ <resources/>
+ <connections/>
+</ui>
