+++ media/FileVerifier.ico.lic
@@ -1,0 +1,1 @@
+https://icons8.de/license
